boxel-cli: extract findCheckpoint into lib/, gate exact-short scan
Address review feedback on PR #4669:

- Move `findCheckpoint` + `FindResult` out of the commands module into
  `src/lib/find-checkpoint.ts`. The commands module no longer exports
  CLI-internal helpers; tests import the utility directly.

- Gate the exact-short-hash scan on `trimmed.length === 7` so common
  numeric refs like '1' / '2' don't allocate or scan the full
  checkpoint list before falling through to index lookup.

Also relocate the unit test from `tests/commands/` to `tests/lib/` to
mirror the source layout.
